Improved LibraryBrowser app. Added helper batch file setting up web server on XP. Updated bootfile. This is 1.0 beta4.
git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS02]$/IronScheme;C28216

+++ IronScheme/IronScheme.LibraryBrowser/Program.cs
@@ -20,18 +20,28 @@ namespace IronScheme.LibraryBrowser
         {
           UseShellExecute = false,
           CreateNoWindow = true,
-          FileName = "IronScheme.WebServer.exe"
+          FileName = "IronScheme.WebServer.exe",
+          RedirectStandardError = true
+          
         }
       };
 
       if (ps.Start())
       {
         Thread.Sleep(1000);
-        Application.EnableVisualStyles();
-        Application.SetCompatibleTextRenderingDefault(false);
-        Application.Run(new MainForm());
 
-        ps.Kill();
+        if (!ps.HasExited)
+        {
+          Application.EnableVisualStyles();
+          Application.SetCompatibleTextRenderingDefault(false);
+          Application.Run(new MainForm());
+
+          ps.Kill();
+        }
+        else
+        {
+          MessageBox.Show(ps.StandardError.ReadToEnd(), "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+        }
       }
     }
   }
